+++ plugin-test/maven/spring-boot-example/demo-lib/src/main/java/dev/flikas/LombokPojo.java
@@ -1,0 +1,20 @@
+package dev.flikas;
+
+import lombok.Data;
+
+@Data
+public class LombokPojo {
+    /**
+     * µØÖ·
+     */
+    private String address;
+    /**
+     * ¶Ë¿Ú
+     */
+    private int port;
+
+    /**
+     * Öµ
+     */
+    private String value;
+}