+++ CHANGELOG.md
@@ -8,7 +8,16 @@ to [Semantic Versioning](https://semver.org).
 
 ### Added
 
+- Code completion
+  for most
+  of [the value providers](https://docs.spring.io/spring-boot/docs/current/reference/html/configuration-metadata.html#configuration-metadata.manual-hints.value-providers).
+- Auto popup value completion after key insertion.
+- Better setter navigate: from configuration property source to configuration file.
+
 ### Changed
+
+- Performance improvement: It is no longer necessary to load the Spring configuration metadata each time the project is
+  opened.
 
 ### Deprecated